+++ Test/ac-library-csharp.Test/Internal/BarrettTest.cs
@@ -25,7 +25,7 @@ namespace AtCoder.Internal
         }
 
         [Fact]
-        public void BarrettBorder()
+        public void BarrettInt32Border()
         {
             for (uint mod = int.MaxValue; mod >= int.MaxValue - 20; mod--)
             {
@@ -40,16 +40,42 @@ namespace AtCoder.Internal
                 }
                 foreach (var a in v)
                 {
-                    long a2 = a;
+                    ulong a2 = a;
                     bt.Mul(a, bt.Mul(a, a)).Should().Be((uint)(a2 * a2 % mod * a2 % mod));
                     foreach (var b in v)
                     {
-                        long b2 = b;
+                        ulong b2 = b;
                         bt.Mul(a, b).Should().Be((uint)(a2 * b2 % mod));
                     }
                 }
             }
         }
 
+        [Fact]
+        public void BarrettUInt32Border()
+        {
+            for (uint mod = uint.MaxValue; mod >= uint.MaxValue - 20; mod--)
+            {
+                var bt = new Barrett(mod);
+                var v = new List<uint>();
+                for (uint i = 0; i < 10; i++)
+                {
+                    v.Add(i);
+                    v.Add(mod - i);
+                    v.Add(mod / 2 + i);
+                    v.Add(mod / 2 - i);
+                }
+                foreach (var a in v)
+                {
+                    ulong a2 = a;
+                    bt.Mul(a, bt.Mul(a, a)).Should().Be((uint)(a2 * a2 % mod * a2 % mod));
+                    foreach (var b in v)
+                    {
+                        ulong b2 = b;
+                        bt.Mul(a, b).Should().Be((uint)(a2 * b2 % mod));
+                    }
+                }
+            }
+        }
     }
 }